requirements.txt
PyQt5
appdirs

azure-identity
azure-mgmt-resource
azure-mgmt-iothub
azure-iot-hub

pycairo
PyGObject